ゲームマネージャー消去
＋ギャラリーの追加
