+++ StandZodiacUnity/StandZodiac/Assets/Scenes/CharaRoom.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c3d6624467f52c84facf263ecdce5827
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
