+++ StandZodiacUnity/StandZodiac/Assets/Scenes/Otamesi_Scene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c467244308f4e8146b5e6f1fecffa041
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
