--- StandZodiacUnity/StandZodiac/Assets/Script/GameManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ea449bc37677031458bf94db83cb33bf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
