+++ StandZodiacUnity/StandZodiac/Assets/Script/TimeController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3766d40b3558244bb10d34d00153140
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
