+++ StandZodiacUnity/StandZodiac/Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/CharaSelect.unity
+- path: Assets/Scenes/Gallery.unity
   isLoaded: 1
   isActive: 1
